Sample validation DQM PIDs needs to shift

+++ DQM/include/DQM/SampleValidation.h
@@ -19,7 +19,7 @@ class SampleValidation : public framework::Analyzer {
       : Analyzer(name, process) {}
   virtual void configure(framework::config::Parameters& ps) override;
   virtual void analyze(const framework::Event& event) override;
-  int pdgidLabel(const int pdgid);
+  float pdgidLabel(const int pdgid);
 
  private:
   std::string target_scoring_plane_passname_;
